// Frontend/front/src/app/api/model/productDto.ts
/**
 * CubArt API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductTypeEnum } from './productTypeEnum';
import { UnitOfMeasureEnum } from './unitOfMeasureEnum';
import { ProductSpecificationDto } from './productSpecificationDto';


export interface ProductDto { 
    id: number;
    name: string;
    productType: ProductTypeEnum;
    unitOfMeasure: UnitOfMeasureEnum;
    activeSpecification?: ProductSpecificationDto;
}
export namespace ProductDto {
}